being able to run, but poor performnace
